init5/requirements.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-extra-fields==3.0.2
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
flower==1.0.0
gevent==21.12.0
greenlet==1.1.2
humanize==3.13.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
kombu==5.2.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
redis==4.1.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
